A função 2 já não incomoda a 1, contudo não funciona.
git-svn-id: https://svn.fe.up.pt/repos/lcom1314-t5g02@88 979b2d34-c887-48f1-a8b5-8c534db5e80d

+++ lab4/test4.c
@@ -57,11 +57,14 @@ int test_scan() {
 }
 
 int test_leds(unsigned short n, unsigned short *leds) {
-  int led;
+
+
+  unsigned short led;
   for (led = 0; led < n; led++){
 	  printf("Toggling LED number %d in array.\n",leds[led]);
 	  int i = leds[led];
 	  keyboard_toggle_led(i);
 	  sleep(1);
   }
+
 }
